requirements.txt
paho-mqtt
python-systemd
